backend/requirements.txt
# Backend WasteVision AI – Python 3.10, 3.11 ou 3.12 requis (TensorFlow ne supporte pas 3.14)
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
tensorflow>=2.15.0
opencv-python-headless>=4.8.0
pydantic>=2.0.0
pydantic_core>=2.0.0
# Optional: for YOLOv8 multi-object detection (place weights in data/weights/yolov8_waste.pt)
ultralytics>=8.0.0
